hot fix:   - moved app_layout package outside of counters, since it will deal with other apps as well
